[pose_estimator] Improve PC filtering using agg. clustering
This particularly addresses cases when the TB is positioned diagonally
on a flat surface and most of the points hitting the surface are
erroneously considered inliers.
The added agglomerative clustering enables identifying the cluster of
points that fall on the surface of the TB and removing such outliers,
which tend to cause errors in orientation estimation.