[cont_cnn_detector] Refactor for blocking UDP triggers
Instead of the main while loop running through continuously and the
detection only being activated at an iteration if a trigger (ROS)
message is received, the iteration is now blocked until a trigger
message is received.
This is in preparation for integrating the UDP message triggers, which
necessarily block while listening, unlike the ROS subscribers.